Fix several invalid memory accesses (thanks Valgrind). Fix 2 memory leaks: Map leaked *m_mtm and Critter_Bob_Descr leaked programs (thanks Valgrind). Change a bunch of members, parameters and returns from const char * to const std::string &. Change Tribe_Descr(std::string name) to Tribe_Descr(const std::string & name). Shorten names of enumeration constants in Player. Compiler warning fixes and const fixes.
git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@1722 37b2a8de-5219-0410-9f54-a31bc463ab9c

+++ src/immovable.cc
@@ -830,7 +830,7 @@ Cleanup
 PlayerImmovable::~PlayerImmovable()
 {
 	if (m_workers.size())
-		log("Building::~Building: %i workers left!\n", (int)m_workers.size());
+		log("Building::~Building: %i workers left!\n", m_workers.size());
 }
 
 /*
